+++ CHANGELOG.md
@@ -27,9 +27,9 @@
   By default, using the handler will skip the current production.
   A combination can be specified, such as `['handler', 'null']`.
 
-* object.transform now supports an `on_missing` option with
-  possible values `null`, `skip` or `raise`. Default behavior is
-  to raise an error if the input object lacks a key.
+* object.transform and object.select now support an `on_missing`
+  option with possible values `null`, `skip` or `raise`. Default
+  behavior is to raise an error if the input object lacks a key.
 
 ## 0.2.0 - 2022-05-05